backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
elevenlabs
replicate
moviepy
firebase-admin
supabase
imageio-ffmpeg
python-dotenv

